LPD-57956 Fix tests
