+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/test/java/com/liferay/portal/search/elasticsearch7/internal/index/IndexFactoryFixture.java
@@ -30,9 +30,9 @@ import org.osgi.framework.FrameworkUtil;
 /**
  * @author Adam Brandizzi
  */
-public class CompanyIndexFactoryFixture {
+public class IndexFactoryFixture {
 
-	public CompanyIndexFactoryFixture(
+	public IndexFactoryFixture(
 		ElasticsearchClientResolver elasticsearchClientResolver,
 		String indexName) {
 
@@ -52,49 +52,23 @@ public class CompanyIndexFactoryFixture {
 	}
 
 	public void createIndices() {
-		CompanyIndexFactory companyIndexFactory = getCompanyIndexFactory();
+		IndexFactory indexFactory = getIndexFactory();
 
 		RestHighLevelClient restHighLevelClient =
 			_elasticsearchClientResolver.getRestHighLevelClient();
 
-		companyIndexFactory.initializeIndex(
+		indexFactory.initializeIndex(
 			RandomTestUtil.randomLong(), restHighLevelClient.indices());
 	}
 
 	public void deleteIndices() {
-		CompanyIndexFactory companyIndexFactory = getCompanyIndexFactory();
+		IndexFactory indexFactory = getIndexFactory();
 
 		RestHighLevelClient restHighLevelClient =
 			_elasticsearchClientResolver.getRestHighLevelClient();
 
-		companyIndexFactory.deleteIndex(
+		indexFactory.deleteIndex(
 			RandomTestUtil.randomLong(), restHighLevelClient.indices());
-	}
-
-	public CompanyIndexFactory getCompanyIndexFactory() {
-		if (_companyIndexFactory != null) {
-			return _companyIndexFactory;
-		}
-
-		_companyIndexFactory = new CompanyIndexFactory();
-
-		ReflectionTestUtil.setFieldValue(
-			_companyIndexFactory, "_companyIndexHelper",
-			getCompanyIndexHelper());
-		ReflectionTestUtil.setFieldValue(
-			_companyIndexFactory, "_companyLocalService",
-			Mockito.mock(CompanyLocalService.class));
-		ReflectionTestUtil.setFieldValue(
-			_companyIndexFactory, "_elasticsearchConfigurationWrapper",
-			createElasticsearchConfigurationWrapper());
-		ReflectionTestUtil.setFieldValue(
-			_companyIndexFactory, "_elasticsearchConnectionManager",
-			_elasticsearchConnectionManager);
-
-		ReflectionTestUtil.invoke(
-			_companyIndexFactory, "activate", new Class<?>[0]);
-
-		return _companyIndexFactory;
 	}
 
 	public CompanyIndexHelper getCompanyIndexHelper() {
@@ -130,6 +104,19 @@ public class CompanyIndexFactoryFixture {
 		return _companyIndexHelper;
 	}
 
+	public IndexFactory getIndexFactory() {
+		if (_indexFactory != null) {
+			return _indexFactory;
+		}
+
+		_indexFactory = new IndexFactory(
+			getCompanyIndexHelper(), Mockito.mock(CompanyLocalService.class),
+			createElasticsearchConfigurationWrapper(),
+			_elasticsearchConnectionManager);
+
+		return _indexFactory;
+	}
+
 	public String getIndexName() {
 		IndexName indexName = new IndexName(_indexName);
 
@@ -137,11 +124,10 @@ public class CompanyIndexFactoryFixture {
 	}
 
 	public void tearDown() {
-		if (_companyIndexFactory != null) {
-			ReflectionTestUtil.invoke(
-				_companyIndexFactory, "deactivate", new Class<?>[0]);
+		if (_indexFactory != null) {
+			_indexFactory.close();
 
-			_companyIndexFactory = null;
+			_indexFactory = null;
 		}
 
 		if (_companyIndexHelper != null) {
@@ -172,7 +158,7 @@ public class CompanyIndexFactoryFixture {
 
 		@Override
 		public String getIndexName(long companyId) {
-			return CompanyIndexFactoryFixture.this.getIndexName();
+			return IndexFactoryFixture.this.getIndexName();
 		}
 
 		@Override
@@ -210,12 +196,12 @@ public class CompanyIndexFactoryFixture {
 		return searchEngineInformation;
 	}
 
-	private CompanyIndexFactory _companyIndexFactory;
 	private CompanyIndexHelper _companyIndexHelper;
 	private final ElasticsearchClientResolver _elasticsearchClientResolver;
 	private final ElasticsearchConnectionManager
 		_elasticsearchConnectionManager;
 	private MockedStatic<FrameworkUtil> _frameworkUtilMockedStatic;
+	private IndexFactory _indexFactory;
 	private final String _indexName;
 
 }
