+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/test/resources/com/liferay/portal/search/elasticsearch7/internal/index/IndexFactoryTest-legacyOverrideTypeMappings.json
