+++ modules/apps/portal-search-opensearch2/portal-search-opensearch2-impl/src/test/java/com/liferay/portal/search/opensearch2/internal/index/CompanyIdIndexNameBuilderTest.java
@@ -78,11 +78,10 @@ public class CompanyIdIndexNameBuilderTest extends BaseOpenSearchTestCase {
 
 	@After
 	public void tearDown() throws Exception {
-		if (_companyIndexFactory != null) {
-			ReflectionTestUtil.invoke(
-				_companyIndexFactory, "deactivate", new Class<?>[0]);
+		if (_indexFactory != null) {
+			_indexFactory.close();
 
-			_companyIndexFactory = null;
+			_indexFactory = null;
 		}
 
 		if (_companyIndexHelper != null) {
@@ -183,28 +182,14 @@ public class CompanyIdIndexNameBuilderTest extends BaseOpenSearchTestCase {
 			new Class<?>[] {BundleContext.class},
 			SystemBundleUtil.getBundleContext());
 
-		_companyIndexFactory = new CompanyIndexFactory();
-
-		ReflectionTestUtil.setFieldValue(
-			_companyIndexFactory, "_companyLocalService",
-			Mockito.mock(CompanyLocalService.class));
-		ReflectionTestUtil.setFieldValue(
-			_companyIndexFactory, "_companyIndexHelper", _companyIndexHelper);
-		ReflectionTestUtil.setFieldValue(
-			_companyIndexFactory, "_openSearchConfigurationWrapper",
-			openSearchConfigurationWrapper);
-		ReflectionTestUtil.setFieldValue(
-			_companyIndexFactory, "_openSearchConnectionManager",
-			openSearchConnectionManager);
-
-		ReflectionTestUtil.invoke(
-			_companyIndexFactory, "activate", new Class<?>[0]);
+		_indexFactory = new IndexFactory(
+			_companyIndexHelper, Mockito.mock(CompanyLocalService.class),
+			openSearchConfigurationWrapper, openSearchConnectionManager);
 
 		OpenSearchClient openSearchClient =
 			openSearchConnectionManager.getOpenSearchClient();
 
-		_companyIndexFactory.initializeIndex(
-			companyId, openSearchClient.indices());
+		_indexFactory.initializeIndex(companyId, openSearchClient.indices());
 	}
 
 	private void _assertIndexNamePrefix(
@@ -289,7 +274,7 @@ public class CompanyIdIndexNameBuilderTest extends BaseOpenSearchTestCase {
 	private static final MockedStatic<FrameworkUtil>
 		_frameworkUtilMockedStatic = Mockito.mockStatic(FrameworkUtil.class);
 
-	private CompanyIndexFactory _companyIndexFactory;
 	private CompanyIndexHelper _companyIndexHelper;
+	private IndexFactory _indexFactory;
 
 }
