+++ modules/apps/portal-search-opensearch2/portal-search-opensearch2-impl/src/test/resources/com/liferay/portal/search/opensearch2/internal/index/IndexFactoryTest-overrideTypeMappings.json
